@@ -5,7 +5,6 @@ requires = ["poetry>=0.12"]
 [tool.poetry]
 authors = ["Anthony Lukach <anthonylukach@gmail.com>"]
 description = "Replace Pydantic's builtin Secret Support with a configuration provider that loads parameters from AWS Systems Manager Parameter Store."
-homepage = ""
 license = "MIT"
 maintainers = ["Anthony Lukach <anthonylukach@gmail.com>"]
 name = "pydantic-ssm-settings"
@@ -15,8 +14,9 @@ version = "0.2.4"
 
 [tool.poetry.dependencies]
 "boto3" = "^1.21.45"
-pydantic = "^1.6.2"
-python = "^3.7"
+pydantic = "^2.5.2"
+python = "^3.9"
+pydantic-settings = ">=2.1.0"
 
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
