@@ -1,4 +1,5 @@
-from .settings import AwsSsmSourceConfig
+from .settings import AwsSsmSourceConfig, SsmSettingsConfigDict
+from .source import AwsSsmSettingsSource
 
-__all__ = ("AwsSsmSourceConfig",)
+__all__ = ("AwsSsmSourceConfig", "SsmSettingsConfigDict", "AwsSsmSettingsSource")
 __version__ = "0.2.4"
